requirements.test.txt
google-generativeai==0.7.2
langchain==0.3.1
langchain-community==0.3.1
langchain-chroma==0.1.4
langchain-google-genai==2.0.0
unstructured==0.15.13
unstructured[pdf]
chromadb==0.5.11
PyMuPDF==1.24.11
aiofiles==24.1.0
fastapi==0.115.0
fastapi-limiter==0.1.6
loguru==0.7.2
redis==5.1.1
celery==5.4.0
deepeval==1.3.2
httpx==0.27.2
pytest-asyncio==0.24.0